Refactored mcmc module into a package
